Re-organized code into src dir
