+++ src/bleManager.tsx
