+++ src/components/DeviceSettingsList.tsx
