+++ src/screens/DataCollectionScreen.tsx
